+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.3.0] - 2026-07-09
+
+_Stable release based on [5.3.0-rc.1]._
+
+## [5.3.0-rc.1] - 2026-07-09
+
+### Added
+
+- Cap CSV exports at a configurable row limit ~ `exportLimit` parameter or `model-browser.export_limit` config, default 1500.
+
 ## [5.2.1] - 2026-07-08
 
 ### Fixed
@@ -756,6 +766,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[5.3.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.2.1...v5.3.0
+[5.3.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v5.2.1
 [5.2.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.2.0...v5.2.1
 [5.2.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v5.1.1...v5.2.0
 [5.2.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v5.1.1